+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/repository/restaurantmenu/RestaurantMenuLikeRepositoryImpl.java
@@ -1,4 +1,4 @@
-package ProjectDoge.StudentSoup.repository.restaurant;
+package ProjectDoge.StudentSoup.repository.restaurantmenu;
 
 import ProjectDoge.StudentSoup.entity.restaurant.RestaurantMenuLike;
 import com.querydsl.jpa.impl.JPAQueryFactory;
