+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/service/restaurantmenu/RestaurantMenuFindService.java
@@ -3,7 +3,7 @@ package ProjectDoge.StudentSoup.service.restaurantmenu;
 import ProjectDoge.StudentSoup.entity.restaurant.RestaurantMenu;
 import ProjectDoge.StudentSoup.exception.restaurant.RestaurantMenuNotFoundException;
 import ProjectDoge.StudentSoup.exception.restaurant.RestaurantMenuNotSentException;
-import ProjectDoge.StudentSoup.repository.restaurant.RestaurantMenuRepository;
+import ProjectDoge.StudentSoup.repository.restaurantmenu.RestaurantMenuRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
